refactor: default locale file
