+++ components/form/FormItemLabel.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import type { ColProps } from '../grid/col';
 import Col from '../grid/col';
 import { useLocaleReceiver } from '../locale-provider/LocaleReceiver';
-import defaultLocale from '../locale/default';
+import defaultLocale from '../locale/en_US';
 import type { TooltipProps } from '../tooltip';
 import Tooltip from '../tooltip';
 import type { FormContextProps } from './context';
