--- components/locale-provider/default.tsx
@@ -1,3 +1,0 @@
-import locale from '../locale/default';
-
-export default locale;
