+++ components/transfer/ListItem.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import type { KeyWiseTransferItem } from '.';
 import Checkbox from '../checkbox';
 import LocaleReceiver from '../locale-provider/LocaleReceiver';
-import defaultLocale from '../locale/default';
+import defaultLocale from '../locale/en_US';
 import TransButton from '../_util/transButton';
 
 type ListItemProps<RecordType> = {
